+++ src/components/cui-column/cui-column.spec.ts
@@ -1,4 +1,4 @@
-import { cColumn } from './cui-column';
+import { cColumn } from './c-column';
 describe('app', () => {
     it('builds', () => {
         expect(new cColumn()).toBeTruthy();
