+++ src/components/cui-container/cui-container.spec.ts
@@ -1,4 +1,4 @@
-import { cContainer } from './cui-container';
+import { cContainer } from './c-container';
 describe('app', () => {
     it('builds', () => {
         expect(new cContainer()).toBeTruthy();
