+++ src/components/cui-content/cui-content.spec.ts
@@ -1,4 +1,4 @@
-import { cContent } from './cui-content';
+import { cContent } from './c-content';
 describe('app-root', () => {
     it('builds', () => {
         expect(new cContent()).toBeTruthy();
