+++ src/components/cui-field/cui-field.spec.ts
@@ -1,4 +1,4 @@
-import { cField } from './cui-field';
+import { cField } from './c-field';
 describe('app', () => {
     it('builds', () => {
         expect(new cField()).toBeTruthy();
