+++ src/components/cui-header/cui-header.spec.ts
@@ -1,4 +1,4 @@
-import { cHeader } from './cui-header';
+import { cHeader } from './c-header';
 describe('app', () => {
     it('builds', () => {
         expect(new cHeader()).toBeTruthy();
