+++ src/components/cui-header/cui-header.tsx
@@ -1,6 +1,6 @@
 import { Component, Prop, State, Method, Watch } from '@stencil/core';
 import { store } from '../../store';
-import * as style from '../../themes.built/cui-header';
+import * as style from '../../themes.built/c-header';
 
 @Component({
   tag: 'c-header',
