+++ src/components/cui-icon/cui-icon.spec.ts
@@ -1,4 +1,4 @@
-import { cIcon } from './cui-icon';
+import { cIcon } from './c-icon';
 describe('app', () => {
     it('builds', () => {
         expect(new cIcon()).toBeTruthy();
