+++ src/components/cui-list/cui-list.spec.ts
@@ -1,4 +1,4 @@
-import { cList } from './cui-list';
+import { cList } from './c-list';
 describe('app', () => {
     it('builds', () => {
         expect(new cList()).toBeTruthy();
