+++ src/components/cui-navigation/cui-navigation.spec.ts
@@ -1,4 +1,4 @@
-import { cNavigation } from './cui-navigation';
+import { cNavigation } from './c-navigation';
 describe('app', () => {
     it('builds', () => {
         expect(new cNavigation()).toBeTruthy();
