+++ src/components/cui-row/cui-row.spec.ts
@@ -1,4 +1,4 @@
-import { cRow } from './cui-row';
+import { cRow } from './c-row';
 describe('app', () => {
     it('builds', () => {
         expect(new cRow()).toBeTruthy();
