+++ src/components/cui-theme/cui-theme.spec.ts
@@ -1,4 +1,4 @@
-import { cTheme } from './cui-theme';
+import { cTheme } from './c-theme';
 describe('app-root', () => {
     it('builds', () => {
         expect(new cTheme()).toBeTruthy();
